refactor: extract styles and move components into new folders
